requirements.txt
discord.py>=2.0.0
google-genai>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytest>=7.0.0
flake8>=5.0
redis>=4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
requests>=2.25.0
